+++ programs/tally-clob/src/state/sub_market_portfolio.rs
@@ -25,7 +25,7 @@ impl SubMarketPortfolio {
         }
     }
          
-    pub fn add_to_portfolio(&mut self, choice_id: &u64, shares: f64) -> Result<&Self> {
+    pub fn add_to_portfolio(&mut self, choice_id: &u64, shares: u128) -> Result<&Self> {
         self
             .get_choice_market_portfolio(choice_id)?
             .add_to_portfolio(shares)?;
@@ -34,7 +34,7 @@ impl SubMarketPortfolio {
         Ok(self)
     }
 
-    pub fn sell_from_portfolio(&mut self, choice_id: &u64, shares: f64) -> Result<&Self> {
+    pub fn sell_from_portfolio(&mut self, choice_id: &u64, shares: u128) -> Result<&Self> {
         self
             .get_choice_market_portfolio(choice_id)?
             .withdraw_from_portfolio(shares)?;
@@ -42,7 +42,7 @@ impl SubMarketPortfolio {
         Ok(self)
     }
 
-    pub fn get_choice_shares(&mut self, choice_id: &u64) -> Result<f64> {
+    pub fn get_choice_shares(&mut self, choice_id: &u64) -> Result<u128> {
         let choice_shares = self
             .get_choice_market_portfolio(choice_id)?
             .shares;
